Split up footer component into different file
